Patch bug where loading certain galleries causes image in the middle to appear.
Loading galleries of Monetary, Are You, and Cape Ann sometimes initially shows the middle of the slideshow on desktop. Something about the scroll-snap-type rule initially causes this on both Safari and Chrome.

Don't add the scroll-snap-type CSS until a little after the page has loaded to give some time for the images to at least start loading? It seems to prevent the slideshow from starting at a random position.

Using a timeout of 0 did not fix the issue.